adde and styled segment filters
