+++ modules/manifold/src/components/ui/segment-filter-panel/index.js
@@ -1,1 +1,1 @@
-export {default as SegmentFilterPanel} from './segment-filter-panel';
+export {default as SegmentFiltersPanel} from './segment-panel-list';
