+++ modules/manifold/src/constants/index.js
@@ -1,3 +1,4 @@
 export * from './constants';
 export {HELP_PAGES} from './help-pages';
 export {default as THEME} from './theme';
+export * from './types';
