--- modules/mlvis-common/src/ui/help-dialog/index.js
@@ -1,3 +1,0 @@
-import HelpDialog from './help-dialog';
-
-export default HelpDialog;
