--- modules/mlvis-common/src/ui/legend-group/index.js
@@ -1,4 +1,0 @@
-// @noflow
-import LegendGroup from './legend-group';
-
-export default LegendGroup;
