--- modules/segment-filters/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@mlvis/segment-filters",
-  "version": "1.0.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "description": "Segment Filter component of the Manifold",
-  "license": "Apache-2.0",
-  "author": "Firenze11 <lezhi.li@uber.com>",
-  "main": "dist/esm/index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/uber/manifold.git"
-  },
-  "keywords": [
-    "ml",
-    "visualization",
-    "model agnostic",
-    "model debugging",
-    "babel",
-    "react",
-    "redux",
-    "es6"
-  ],
-  "dependencies": {
-    "@mlvis/mlvis-common": "^1.0.0",
-    "prop-types": "^15.7.2"
-  },
-  "peerDependencies": {
-    "react": "^16.7.0",
-    "styled-components": "^4.2.0"
-  }
-}
