--- modules/segment-filters/src/index.js
@@ -1,8 +1,0 @@
-// @noflow
-export {default as SegmentFilters} from './segment-panel-list';
-
-// TODO:
-// 1. make segment filters width responsive - P1
-// 2. do not use fixed padding left for the confirm button - P1
-// 3. hover over distribution to show domain and value - P1
-// 4. allow overwriting background styles - P1
